+++ CHANGELOG.md
@@ -2,6 +2,159 @@
 
 ## 2021
 
+### 07-28 v2.29.14
+
+- 修复一些 saucenao 错误原因的判断 ([#232](../../issues/232))
+
+### 07-24 v2.29.13
+
+- ascii2d 出现“first byte timeout”错误时进行重试 ([#227](../../issues/227))
+
+### 07-01 v2.29.12
+
+- 修复 setu 不能 r18 的问题
+
+### 06-26 v2.29.11
+
+- setu 使用 API v2，keyword 支持使用 & 和 |（详见 wiki），且不再需要 API Key
+- 修复特定情况下 nhentai 搜不到本子的问题
+- QQ OCR 增加重试机制，避免偶然抽风
+- 完善 saucenao 中图源、标题、作者的获取
+- 配置项变更
+  - A `bot.replys.setuNotFind` 没有符合条件 setu 时的回复
+
+### 06-05 v2.29.10
+
+- 修复 ascii2d 搜索失败仍会缓存结果的问题 ([#202](../../issues/202))
+- 修复特定情况下 OCR 抛出异常的问题 ([#203](../../issues/203))
+
+### 05-29 v2.29.8
+
+- 提升本子搜索成功率
+- 修复一些问题
+
+### 05-29 v2.29.7
+
+- 对 saucenao 的 pixiv 结果进行权重处理 ([#199](../../issues/199))
+
+### 05-29 v2.29.6
+
+- 修正搜图和 setu 的配额机制，防止绕过限制 ([#200](../../issues/200))
+- 语言库支持使用 `[CQ:delete]` 撤回对方消息
+- 修正方舟公招计算器结果图中英文和数字的文字基线
+
+### 05-28 v2.29.5
+
+- 修复可能出现 saucenao 搜图 404 的问题 ([#198](../../issues/198))
+
+### 05-27 v2.29.4
+
+- 修复群聊 saucenao 搜索结果异常的问题 ([#195](../../issues/195))
+- 修复部分情况下会缓存失败搜索结果的问题
+
+### 05-25 v2.29.3
+
+- 修复部分情况下可能出现文件不存在错误的问题（不影响程序正常运行）
+- whatanime 发送预览视频不再需要依赖 ffmpeg
+- 改进方舟公招计算器的识别逻辑
+- 默认 OCR 服务变更为为 qq
+- `bot.setu.pximgProxy` 和 `bot.setu.sendPximgProxys` 新增支持一些占位符
+- 配置项变更
+  - M `bot.ocr.use` 默认值 `"ocr.space"` -> `"qq"`
+  - M `bot.akhr.ocr` 默认值 `"ocr.space"` -> `"qq"`
+
+### 05-23 v2.29.2
+
+- 改进方舟公招计算器生成图片的效果 ([#193](../../issues/193))
+- 新增 `autoUpdateConfig` 配置，可自动按照 `config.default.json` 来更新 `config.json`
+- 配置项变更
+  - A `autoUpdateConfig`
+
+### 05-23 v2.29.1
+
+修复找不到模块问题，若先前更新过 v2.29.0，更新到最新版时请按以下步骤操作
+
+```bash
+git reset v2.29.0 --hard
+git pull # 可能会报错，不用管
+git reset v2.29.1 --hard
+git pull
+```
+
+### 05-23 v2.29.0
+
+- 修复了 `bot.setu.r18OnlyUrl` 导致非 r18 图也只发 url 的问题 ([#182](../../issues/182))
+- 更新了 trace.moe API，解决了无法发送预览视频的问题
+- `bot.setu.r18OnlyUrl` 分开群聊、私聊、临时会话（⚠️需要更新配置） ([#183](../../issues/183))
+- 定时提醒可通过添加 `--origin` 参数使内容不被 CQ 转义，见 [wiki](../../wiki/%E9%99%84%E5%8A%A0%E5%8A%9F%E8%83%BD#%E9%98%B2%E6%AD%A2-cq-%E8%BD%AC%E4%B9%89) ([#178](../../issues/178))
+- 可以使用 reminder 定时发送 setu，见 [wiki](../../wiki/%E9%99%84%E5%8A%A0%E5%8A%9F%E8%83%BD#setu-1) ([#174](../../issues/174))
+- 配置项变更
+  - M `bot.whatanimeHost` 默认值 `"trace.moe"` -> `"api.trace.moe"`（旧值在读取时会被自动替换为新值）
+  - M `bot.setu.r18OnlyUrl` 默认值 `false` -> `{ "private": false, "group": false, "temp": false }`，如先前修改为 `true` 请更新配置，否则会变为新默认值
+
+### 05-19 v2.28.5
+
+- 修复了 `bot.setu.pximgServerHost` 设置无效的问题
+- 弃用了一些已经不可用的短链接服务
+
+### 05-17 v2.28.4
+
+- 修复了可能出现 pm2 占用内存过多的问题
+- ⚠️本次更新前请在程序目录下执行 `npx pm2 delete CQPF`，不需要执行 `npm stop`，其他步骤相同
+
+### 05-16 v2.28.3
+
+- 修复“允许通过临时会话私聊发送”相关功能可能无效的问题
+
+### 05-16 v2.28.2
+
+- 修复 pximg 反代服务启动问题
+- pximg 反代服务默认只在本地(`127.0.0.1`)监听，如有需要请更改 `bot.setu.pximgServerHost`
+- 配置项变更
+  - A `bot.setu.pximgServerHost`
+
+### 05-16 v2.28.1
+
+- 修复 `bot.setu.shortenPximgProxy` 默认值
+
+### 05-16 v2.28.0
+
+- [#175](../../pull/175) by @niceRAM
+  - 允许通过临时会话私聊发送搜图结果
+  - 可配置 r18 setu 仅通过私聊发送（默认开启，如不需要请修改配置）
+  - 允许 r18 setu 通过临时会话私聊发送
+  - 发送 setu 链接时可以追加若干个原图镜像链接
+  - 可配置对原图镜像链接做短链接处理（增加 oy.mk 短链接服务 by @Quan666）
+- pximg 反代服务默认使用随机可用端口号（老用户可手动将 `bot.setu.pximgServerPort` 设置为 `0`）
+- 配置项变更
+  - M `bot.setu.pximgServerPort` 默认值 `60233` -> `0`
+  - A `bot.pmSearchResultTemp`
+  - A `bot.setu.sendPximgProxys`
+  - A `bot.setu.shortenPximgProxy`
+  - A `bot.setu.r18OnlyUrl`
+  - A `bot.setu.r18OnlyPrivate`
+  - A `bot.setu.r18OnlyPrivateAllowTemp`
+
+### 05-12 v2.27.1
+
+- 使用方舟公招计算器将不再需要自行安装字体
+
+### 05-11 v2.27.0
+
+- 搜图缓存使用文件缓存代替 sqlite 以节省内存，注意：`data/db.sqlite` 将会被删除
+
+### 05-05 v2.26.1
+
+- 使用 go-cqhttp 的场合下，whatanime 支持发送预览视频的功能需要自行安装 [ffmpeg](https://ffmpeg.org/download.html) 才可以使用
+- 配置项变更
+  - M `bot.whatanimeSendVideo` 默认变更为 `false`
+
+### 05-05 v2.26.0
+
+- whatanime 支持发送预览视频
+- 配置项变更
+  - A `bot.whatanimeSendVideo`
+
 ### 04-25 v2.25.0
 
 - 定时提醒支持重载了（方便手动修改 `data/rmd.json`）